fft yes 1

+++ fft.h
@@ -133,7 +133,7 @@ public:
     struct T
     {
         int i;
-        T(int _i = 0):i(_i){}
+        T(LL _i = 0):i(_i){}
         friend T operator*(const T &a, const T &b)
         {
             return T(1LL * a.i * b.i % P);
